+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - added new feature inactive
 - autoInit on documentReady
 - fixed breakpoints ( bug: started before init, when addHorizontalBreakPoint method used )
+- fixed time-based ( bug: shift is not a function )
 
 ## 3.0.0 Early Bird
 - library custom builds with missing method warnings